LayoutTests/fast/exclusions/shape-inside/shape-inside-polygon-padding-003-expected.txt

XXX

XX

XXX
PASS shapeInsideRect('a').top is 50
PASS shapeInsideRect('a').width is 150
PASS shapeInsideRect('b').top is 150
PASS shapeInsideRect('b').left is within 1 of 250
PASS shapeInsideRect('b').width is 100
PASS shapeInsideRect('c').top is 250
PASS shapeInsideRect('c').width is 150
